Wrapping-up documentation and code for DL-Centric experiments

+++ src/FF_proj/source/FullyConnectedNetwork.cc
@@ -10,12 +10,12 @@
 
 GET_V_TABLE(FullyConnectedNetwork)
 
-double sigmoid(double x) { return 1 / (1 + exp(-x)); }
+float sigmoid(float x) { return 1 / (1 + exp(-x)); }
 
-double relu(double x) { return std::max(0.0, x); }
+float relu(float x) { return std::max(0.0f, x); }
 
-double outLabel(double x) {
-    double threshold = 0.5;
+float outLabel(float x) {
+    float threshold = 0.5;
     if (x > threshold) {
         return 1;
     } else {
@@ -23,7 +23,7 @@ double outLabel(double x) {
     }
 }
 
-void randomGen(Handle<Vector<double>> vec) {
+void randomGen(Handle<Vector<float>> vec) {
 
     std::random_device rd;
